+++ js/app.js
@@ -1,10 +1,9 @@
 import { db } from './db.js';
 import * as workout from './views/workout.js';
-import * as routines from './views/routines.js';
 import * as statistics from './views/statistics.js';
 import * as profile from './views/profile.js';
 
-const views = { workout, routines, statistics, profile };
+const views = { workout, statistics, profile };
 
 const viewContainer = document.getElementById('view-container');
 const navButtons = document.querySelectorAll('.nav-btn');